+++ core/src/main/java/com/backblaze/b2/client/structures/B2CreateBucketRequestReal.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021, Backblaze Inc. All Rights Reserved.
+ * Copyright 2023, Backblaze Inc. All Rights Reserved.
  * License https://www.backblaze.com/using_b2_code.html
  */
 package com.backblaze.b2.client.structures;
@@ -46,8 +46,10 @@ public class B2CreateBucketRequestReal {
     @B2Json.optional
     private final B2BucketReplicationConfiguration replicationConfiguration;
 
-    @B2Json.constructor(params = "accountId,bucketName,bucketType,bucketInfo,corsRules,lifecycleRules,fileLockEnabled,"+
-            "defaultServerSideEncryption, replicationConfiguration")
+    @B2Json.optional
+    private final List<B2EventNotificationRule> eventNotificationRules;
+
+    @B2Json.constructor
     private B2CreateBucketRequestReal(String accountId,
                                       String bucketName,
                                       String bucketType,
@@ -56,7 +58,8 @@ public class B2CreateBucketRequestReal {
                                       List<B2LifecycleRule> lifecycleRules,
                                       boolean fileLockEnabled,
                                       B2BucketServerSideEncryption defaultServerSideEncryption,
-                                      B2BucketReplicationConfiguration replicationConfiguration) {
+                                      B2BucketReplicationConfiguration replicationConfiguration,
+                                      List<B2EventNotificationRule> eventNotificationRules) {
         this.accountId = accountId;
         this.bucketName = bucketName;
         this.bucketType = bucketType;
@@ -66,6 +69,7 @@ public class B2CreateBucketRequestReal {
         this.fileLockEnabled = fileLockEnabled;
         this.defaultServerSideEncryption = defaultServerSideEncryption;
         this.replicationConfiguration = replicationConfiguration;
+        this.eventNotificationRules = eventNotificationRules;
     }
 
     public B2CreateBucketRequestReal(String accountId,
@@ -78,8 +82,8 @@ public class B2CreateBucketRequestReal {
                 mostOfRequest.getLifecycleRules(),
                 mostOfRequest.isFileLockEnabled(),
                 mostOfRequest.getDefaultServerSideEncryption(),
-                mostOfRequest.getReplicationConfiguration()
-        );
+                mostOfRequest.getReplicationConfiguration(),
+                mostOfRequest.getEventNotificationRules());
     }
 
     @Override
@@ -95,7 +99,8 @@ public class B2CreateBucketRequestReal {
                 Objects.equals(lifecycleRules, that.lifecycleRules) &&
                 Objects.equals(fileLockEnabled, that.fileLockEnabled) &&
                 Objects.equals(defaultServerSideEncryption, that.defaultServerSideEncryption) &&
-                Objects.equals(replicationConfiguration, that.replicationConfiguration);
+                Objects.equals(replicationConfiguration, that.replicationConfiguration) &&
+                Objects.equals(eventNotificationRules, that.eventNotificationRules);
     }
 
     @Override
@@ -109,7 +114,8 @@ public class B2CreateBucketRequestReal {
                 lifecycleRules,
                 fileLockEnabled,
                 defaultServerSideEncryption,
-                replicationConfiguration
+                replicationConfiguration,
+                eventNotificationRules
         );
     }
 
@@ -125,6 +131,7 @@ public class B2CreateBucketRequestReal {
                 ", fileLockEnabled=" + fileLockEnabled +
                 ", defaultServerSideEncryption=" + defaultServerSideEncryption +
                 ", replicationConfiguration=" + replicationConfiguration +
+                ", eventNotificationRules=" + eventNotificationRules +
                 '}';
     }
 }