+++ core/src/test/java/com/backblaze/b2/client/B2TestHelpers.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021, Backblaze Inc. All Rights Reserved.
+ * Copyright 2023, Backblaze Inc. All Rights Reserved.
  * License https://www.backblaze.com/using_b2_code.html
  */
 package com.backblaze.b2.client;
@@ -172,6 +172,7 @@ public class B2TestHelpers {
                 B2Collections.listOf(makeCorsRule()),
                 B2Collections.listOf(makeLifecycleRule(i)),
                 makeBucketOrApplicationKeyOptions(),
+                null,
                 null,
                 null,
                 null,