+++ core/src/test/java/com/backblaze/b2/client/structures/B2BucketTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021, Backblaze Inc. All Rights Reserved.
+ * Copyright 2023, Backblaze Inc. All Rights Reserved.
  * License https://www.backblaze.com/using_b2_code.html
  */
 
@@ -19,6 +19,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.TreeMap;
+import java.util.TreeSet;
 
 import static com.backblaze.b2.client.B2TestHelpers.bucketId;
 import static com.backblaze.b2.util.B2Collections.listOf;
@@ -37,6 +38,8 @@ public class B2BucketTest extends B2BaseTest {
     private final Set<String> optionsSet;
     private final List<B2ReplicationRule> replicationRules;
     private final Map<String, String> sourceToDestinationKeyMapping;
+
+    private final List<B2EventNotificationRule> eventNotificationRules;
 
     @Rule
     public ExpectedException thrown = ExpectedException.none();
@@ -79,6 +82,17 @@ public class B2BucketTest extends B2BaseTest {
         sourceToDestinationKeyMapping.put(
                 "456a0b9a8a7a6a50000fc614e", "555a0a1a2a3a4a70000bc929a"
         );
+
+        eventNotificationRules = listOf(
+                new B2EventNotificationRule(
+                        "ruleName",
+                        new TreeSet<>(listOf("b2:ObjectCreated:Copy")),
+                        "",
+                        new B2WebhookConfiguration("https://www.example.com"),
+                        true,
+                        ""
+                )
+        );
     }
 
     @Test
@@ -92,6 +106,7 @@ public class B2BucketTest extends B2BaseTest {
                 b2CorsRules,
                 lifecycleRules,
                 optionsSet,
+                new B2AuthorizationFilteredResponseField<>(false, null),
                 new B2AuthorizationFilteredResponseField<>(false, null),
                 new B2AuthorizationFilteredResponseField<>(false, null),
                 new B2AuthorizationFilteredResponseField<>(false, null),
@@ -127,6 +142,11 @@ public class B2BucketTest extends B2BaseTest {
                                 sourceToDestinationKeyMapping
                         )
                 );
+        final B2AuthorizationFilteredResponseField<List<B2EventNotificationRule>> eventNotificationRulesContainer =
+                new B2AuthorizationFilteredResponseField<>(
+                        true,
+                        eventNotificationRules
+                );
         final B2Bucket bucket = new B2Bucket(
                 ACCOUNT_ID,
                 bucketId(1),
@@ -139,6 +159,7 @@ public class B2BucketTest extends B2BaseTest {
                 bucketFileLockContainer,
                 bucketSseContainer,
                 bucketReplicationConfigurationContainer,
+                eventNotificationRulesContainer,
                 1
         );
         final String bucketJson = B2Json.toJsonOrThrowRuntime(bucket);
@@ -153,6 +174,10 @@ public class B2BucketTest extends B2BaseTest {
                 "  \"bucketId\": \"" + bucketId(1) + "\",\n" +
                 "  \"bucketName\": \"" + BUCKET_NAME + "\",\n" +
                 "  \"defaultServerSideEncryption\": {\n" +
+                "    \"isClientAuthorizedToRead\": false,\n" +
+                "    \"value\": null\n" +
+                "  },\n" +
+                "  \"eventNotificationRules\": {\n" +
                 "    \"isClientAuthorizedToRead\": false,\n" +
                 "    \"value\": null\n" +
                 "  },\n" +
@@ -179,6 +204,7 @@ public class B2BucketTest extends B2BaseTest {
                 null,
                 null,
                 null,
+                new B2AuthorizationFilteredResponseField<>(false, null),
                 new B2AuthorizationFilteredResponseField<>(false, null),
                 new B2AuthorizationFilteredResponseField<>(false, null),
                 new B2AuthorizationFilteredResponseField<>(false, null),
@@ -213,6 +239,11 @@ public class B2BucketTest extends B2BaseTest {
                                 sourceToDestinationKeyMapping
                         )
                 );
+        final B2AuthorizationFilteredResponseField<List<B2EventNotificationRule>> eventNotificationRulesContainer =
+                new B2AuthorizationFilteredResponseField<>(
+                        true,
+                        eventNotificationRules
+                );
         final B2Bucket bucket = new B2Bucket(
                 ACCOUNT_ID,
                 bucketId(1),
@@ -225,6 +256,7 @@ public class B2BucketTest extends B2BaseTest {
                 bucketFileLockContainer,
                 bucketSseContainer,
                 bucketReplicationConfigurationContainer,
+                eventNotificationRulesContainer,
                 1
         );
         // Convert from B2Bucket -> json
@@ -259,6 +291,24 @@ public class B2BucketTest extends B2BaseTest {
                 "      \"algorithm\": \"AES256\",\n" +
                 "      \"mode\": \"SSE-B2\"\n" +
                 "    }\n" +
+                "  },\n" +
+                "  \"eventNotificationRules\": {\n" +
+                "    \"isClientAuthorizedToRead\": true,\n" +
+                "    \"value\": [\n" +
+                "      {\n" +
+                "        \"disabledReason\": \"\",\n" +
+                "        \"eventTypes\": [\n" +
+                "          \"b2:ObjectCreated:Copy\"\n" +
+                "        ],\n" +
+                "        \"isEnabled\": true,\n" +
+                "        \"name\": \"ruleName\",\n" +
+                "        \"objectNamePrefix\": \"\",\n" +
+                "        \"targetConfiguration\": {\n" +
+                "          \"targetType\": \"webhook\",\n" +
+                "          \"url\": \"https://www.example.com\"\n" +
+                "        }\n" +
+                "      }\n" +
+                "    ]\n" +
                 "  },\n" +
                 "  \"fileLockConfiguration\": {\n" +
                 "    \"isClientAuthorizedToRead\": true,\n" +