+++ core/src/test/java/com/backblaze/b2/client/structures/B2EventNotificationRuleTest.java
@@ -1,0 +1,54 @@
+/*
+ * Copyright 2023, Backblaze Inc. All Rights Reserved.
+ * License https://www.backblaze.com/using_b2_code.html
+ */
+package com.backblaze.b2.client.structures;
+
+import com.backblaze.b2.json.B2Json;
+import com.backblaze.b2.util.B2BaseTest;
+import org.junit.Test;
+
+import java.util.TreeSet;
+
+import static org.junit.Assert.assertEquals;
+
+public class B2EventNotificationRuleTest extends B2BaseTest {
+    @Test
+    public void testToJsonAndBack() {
+        final String jsonString = "{\n" +
+                "  \"disabledReason\": \"\",\n" +
+                "  \"eventTypes\": [\n" +
+                "    \"b2:ObjectCreated:Replica\",\n" +
+                "    \"b2:ObjectCreated:Upload\"\n" +
+                "  ],\n" +
+                "  \"isEnabled\": true,\n" +
+                "  \"name\": \"myRule\",\n" +
+                "  \"objectNamePrefix\": \"\",\n" +
+                "  \"targetConfiguration\": {\n" +
+                "    \"targetType\": \"webhook\",\n" +
+                "    \"url\": \"https://www.example.com\"\n" +
+                "  }\n" +
+                "}";
+        final B2EventNotificationRule converted =
+                B2Json.fromJsonOrThrowRuntime(
+                        jsonString,
+                        B2EventNotificationRule.class
+                );
+
+        final TreeSet<String> eventTypes = new TreeSet<>();
+        eventTypes.add("b2:ObjectCreated:Replica");
+        eventTypes.add("b2:ObjectCreated:Upload");
+        final B2EventNotificationRule defaultConfig =
+                new B2EventNotificationRule(
+                        "myRule",
+                        eventTypes,
+                        "",
+                        new B2WebhookConfiguration("https://www.example.com"),
+                        true,
+                        ""
+                );
+        final String convertedJson = B2Json.toJsonOrThrowRuntime(defaultConfig);
+        assertEquals(defaultConfig, converted);
+        assertEquals(jsonString, convertedJson);
+    }
+}